add testcases to support fee splitting for join/exit pool
