+++ pkg/vault/test/ExitPool.test.ts
@@ -22,7 +22,7 @@ import { sharedBeforeEach } from '@balancer-labs/v2-common/sharedBeforeEach';
 describe('Exit Pool', () => {
   let admin: SignerWithAddress, creator: SignerWithAddress, lp: SignerWithAddress;
   let recipient: SignerWithAddress, relayer: SignerWithAddress;
-  let authorizer: Contract, vault: Contract, feesCollector: Contract;
+  let authorizer: Contract, vault: Contract, feesCollector: Contract, poolFeesCollector: Contract;
   let allTokens: TokenList;
 
   const SWAP_FEE_PERCENTAGE = fp(0.1);
@@ -39,6 +39,7 @@ describe('Exit Pool', () => {
     }));
     vault = vault.connect(lp);
     feesCollector = await deployedAt('ProtocolFeesCollector', await vault.getProtocolFeesCollector());
+    poolFeesCollector = await deployedAt('ProtocolFeesCollector', await vault.getPoolFeesCollector());
 
     const action = await actionId(feesCollector, 'setSwapFeePercentage');
     await authorizer.connect(admin).grantPermission(action, admin.address, ANY_ADDRESS);
@@ -74,7 +75,7 @@ describe('Exit Pool', () => {
     }
 
     sharedBeforeEach('deploy & register pool', async () => {
-      pool = await deploy('MockPool', { args: [vault.address, specialization] });
+      pool = await deploy('MockPool_SF_JoinExit', { args: [vault.address, specialization] });
       poolId = await pool.getPoolId();
     });
 
@@ -334,6 +335,7 @@ describe('Exit Pool', () => {
           const fromRelayer = false;
 
           itExitsCorrectlyWithAndWithoutInternalBalance(dueProtocolFeeAmounts, fromRelayer);
+          itExitsCorrectlyDespitePauseAfterFee(dueProtocolFeeAmounts, fromRelayer, false);
         });
       }
 
@@ -447,7 +449,8 @@ describe('Exit Pool', () => {
               [
                 { account: vault, changes: vaultChanges },
                 { account: recipient, changes: recipientChanges },
-                { account: feesCollector, changes: protocolFeesChanges },
+                // { account: feesCollector, changes: protocolFeesChanges },
+                { account: poolFeesCollector, changes: protocolFeesChanges },
               ]
             );
           });
@@ -518,14 +521,14 @@ describe('Exit Pool', () => {
             });
           });
 
-          it('collects protocol fees', async () => {
-            const previousCollectedFees = await feesCollector.getCollectedFeeAmounts(tokens.addresses);
-            await exitPool({ dueProtocolFeeAmounts, fromRelayer, toInternalBalance, signature });
-            const currentCollectedFees = await feesCollector.getCollectedFeeAmounts(tokens.addresses);
+          // it('collects protocol fees', async () => {
+          //   const previousCollectedFees = await feesCollector.getCollectedFeeAmounts(tokens.addresses);
+          //   await exitPool({ dueProtocolFeeAmounts, fromRelayer, toInternalBalance, signature });
+          //   const currentCollectedFees = await feesCollector.getCollectedFeeAmounts(tokens.addresses);
 
-            // Fees from both sources are lumped together.
-            expect(arraySub(currentCollectedFees, previousCollectedFees)).to.deep.equal(dueProtocolFeeAmounts);
-          });
+          //   // Fees from both sources are lumped together.
+          //   expect(arraySub(currentCollectedFees, previousCollectedFees)).to.deep.equal(dueProtocolFeeAmounts);
+          // });
 
           it('exits multiple times', async () => {
             await Promise.all(
@@ -587,6 +590,76 @@ describe('Exit Pool', () => {
           });
         }
       }
+
+      function itExitsCorrectlyDespitePauseAfterFee(
+        dueProtocolFeeAmounts: BigNumberish[],
+        fromRelayer: boolean,
+        toInternalBalance: boolean,
+        signature?: boolean
+      ) {
+        context('when unpaused', () => {
+          itExitsCorrectly(dueProtocolFeeAmounts, fromRelayer, toInternalBalance, signature);
+        });
+
+        context('when paused', () => {
+          sharedBeforeEach('pause', async () => {
+            const action = await actionId(vault, 'setPaused');
+            await authorizer.connect(admin).grantPermission(action, admin.address, ANY_ADDRESS);
+            await vault.connect(admin).setPaused(true);
+          });
+
+          itExitsCorrectly(dueProtocolFeeAmounts, fromRelayer, toInternalBalance, signature);
+        });
+
+        function itExitsCorrectly(
+          dueProtocolFeeAmounts: BigNumberish[],
+          fromRelayer: boolean,
+          toInternalBalance: boolean,
+          signature?: boolean
+        ) {
+          it('update ratio after joining a pool', async () => {
+            // Tokens are sent to the recipient, so the expected change is positive
+            const expectedUserChanges = toInternalBalance ? array(0) : exitAmounts;
+            const recipientChanges = tokens.reduce(
+              (changes, token, i) => ({ ...changes, [token.symbol]: expectedUserChanges[i] }),
+              {}
+            );
+
+            const expectedVaultChanges = toInternalBalance
+              ? dueProtocolFeeAmounts
+              : arrayAdd(exitAmounts, dueProtocolFeeAmounts);
+
+            const vaultChanges = tokens.reduce(
+              // Tokens are sent from the Vault, so the expected change is negative
+              (changes, token, i) => ({ ...changes, [token.symbol]: bn(expectedVaultChanges[i]).mul(-1) }),
+              {}
+            );
+
+            // Tokens are sent to the Protocol Fees, so the expected change is positive
+            const protocolFeesChanges = tokens.reduce(
+              (changes, token, i) => ({ ...changes, [token.symbol]: dueProtocolFeeAmounts[i] }),
+              {}
+            );
+
+            await expectBalanceChange(
+              () => exitPool({ dueProtocolFeeAmounts, fromRelayer, toInternalBalance, signature }),
+              tokens,
+              [
+                { account: vault, changes: vaultChanges },
+                { account: recipient, changes: recipientChanges },
+                // { account: feesCollector, changes: protocolFeesChanges },
+                { account: poolFeesCollector, changes: protocolFeesChanges },
+              ]
+            );
+
+            // assert the ratio after joining pool
+            await tokens.asyncEach(async (token) => {
+                const ratio = await vault.getIndexRatio(poolId, token.address);
+                expect(ratio).to.equal(bn(1e12)); // fee amount= 1e18 => ratio = 1e18 * 1e18 / totalSupply (1e6 * 1e18)
+            });
+          });
+        }
+     }
     });
   }
 });
